Marking feature based on up to date master branch
Functionality of the branch feature/Create-Marks grafted on the up to date master branch.

By clicking on the Mark-button a user can highlight a moment in the recording.
Those marks are saved to the database and maped to the recording via a one-to-many relationship.

The query getRecordingWithMarks returns the EntryEntity and all linked Marks.